requirements.txt
uvicorn
fastapi
gunicorn
motor
pymongo
pymongo[srv]
python-dotenv
python-multipart
jinja2
